Code Agreement Titles

Creates an ID that contains information on the parties to an agreement, the type of agreement, the date and the linkage to other agreements in the dataset.

code_agreements(dataset = NULL, title, date)

Arguments

dataset
dataset name. If provided without a title and date column, the function finds title and date conforming columns in the dataset automatically if available.

title
title column variable. Ideally, title variable should come from a qPackage database/dataset for which strings were standardised with standardise_titles()

date
date column variable Ideally, date variable should come from a qPackage database/dataset for which dates were standardised with standardise_dates()

Value

a character vector with the qIDs

Examples

IEADB <- dplyr::slice_sample(qEnviron::agreements$IEADB, n = 10)
code_agreements(dataset = IEADB)

#> ✔ Title and date conforming columns in dataset automatically found

#> ✔ Coded agreement parties

#> ✔ Coded agreement type

#> ✔ Coded known agreements

#> ✔ Coded agreement dates

#> ✔ Coded acronyms for agreements

#> ✔ Coded agreement linkages

#> 0 entries were not matched at all.
#> There were 0 duplicated IDs.

#> ✔ Please run `vignette('agreements')` for more information.

#>  [1] "FIN-POL[SPE]_1990N" "SCIRWM_1968E20"     "AI04MA_2010E3"     
#>  [4] "TG13TU_1930A"       "CUB-MEX[FSM]_1976A" "CS07IW_1996A"      
#>  [7] "DNK-NOR[DFG]_2001A" "EST-FIN[RGG]_2002A" "PRPDWM_1972A"      
#> [10] "ALB-GRC[TFI]_2003A"

code_agreements(title = IEADB$Title, date = IEADB$Signature)

#> ✔ Coded agreement parties

#> ✔ Coded agreement type

#> ✔ Coded known agreements

#> ✔ Coded agreement dates

#> ✔ Coded acronyms for agreements

#> ✔ Coded agreement linkages

#> 0 entries were not matched at all.
#> There were 0 duplicated IDs.

#> ✔ Please run `vignette('agreements')` for more information.

#>  [1] "FIN-POL[SPE]_1990N" "SCIRWM_1968E20"     "AI04MA_2010E3"     
#>  [4] "TG13TU_1930A"       "CUB-MEX[FSM]_1976A" "CS07IW_1996A"      
#>  [7] "DNK-NOR[DFG]_2001A" "EST-FIN[RGG]_2002A" "PRPDWM_1972A"      
#> [10] "ALB-GRC[TFI]_2003A"