Creates Numerical IDs from Signature Dates

Agreements should have a unique identification number that is meaningful, we condense their signature dates to produce this number.

code_dates(date)

Arguments

date
A date variable

Value

A character vector with condensed dates

Examples

IEADB <- dplyr::slice_sample(qEnviron::agreements$IEADB, n = 10)
code_dates(IEADB$Title)

#>  [1] "Convention To Ban The Importation Into The Forum Island Countries Of Hazardous And Radioactive Wastes And To Control The Transboundary Movement And Management Of Hazardous Wastes Within The South Pacific Re"                                                               
#>  [2] "Agreement Between The Kingdom Of Hungary And The Czechoslovak Republic Concerning The Fixing Of A Uniform General Close Season For Fishing And The Conditions For The Authorization Of Night Fishing In The Czechoslovak Hungarian Frontier Watercou"                         
#>  [3] "Agreement On Economic, Scientific And Technical Cooperation In The Fields Of Agriculture And Food Industry Between Governments Of Hungary And Bel"                                                                                                                            
#>  [4] "Agreement Of Application Of The Agreement On Flood Warning For The Catchment Basin Of The M"                                                                                                                                                                                  
#>  [5] "Amendments Of 1991 To The Technical Annex Of The Protocol To The Convention On LongRange Transboundary Air Pollution Concerning The Control Of Nitrogen Oxides Or Their Transboundary Fl"                                                                                     
#>  [6] "Agreement Between The Government Of The Russian Federation And The Government Of The Kingdom Of Morocco On Cooperation In The Area Of Marine Fishe"                                                                                                                           
#>  [7] "Agreement Between The Government Of The Federal Republic Of Germany And The Government Of The Republic Of Rwanda Concerning Financial Cooperation In 2011/2013 And Concerning Additional Measures For Climate Change Adaptation Under The Special Energy And Climate Fund In "
#>  [8] "Agreement Between The Government Of The United States Of America And The Government Of Canada On East Coast Fishery Resou"                                                                                                                                                    
#>  [9] "Agreement Amending Article 22 Of The Treaty Of The Southern African Development Commu"                                                                                                                                                                                        
#> [10] "Exchange Of Letters Constituting An Agreement Between France And Spain Amending The Arrangement Of 12 July 1958 Relating To Lake La"